Changed so that createServer is mostly built before used in projects
This means that we can export less dependencies.
