+++ examples/server/koa-middlewares.js
@@ -1,4 +1,4 @@
-export default function middlewares(/* config */) {
+export default function middlewares(/* runtimeSettings */) {
     return [function *(next) {
         yield next;
         if (!this.body) {
