+++ extensions/roc-package-web-app/.eslintrc
@@ -14,6 +14,8 @@
         "babel/generator-star-spacing": [2, { before: false, after: true }],
 
         "import/order": [2, { "groups": ["builtin", "external", "internal", "parent", "sibling", "index"], "newlines-between": "always"}],
-        "import/newline-after-import": [2]
+        "import/newline-after-import": [2],
+        "import/no-extraneous-dependencies": [1],
+        "import/no-unresolved": [1]
     }
 }
