+++ extensions/roc-package-web-app/docs/Dependencies.md
@@ -7,62 +7,6 @@ The dependencies that are available in the project.
 __Version__: ~1.16.0  
 __Extension__: roc-package-web-app  
 __Custom resolve function__:  No  
-### [koa](https://www.npmjs.com/package/koa)
-__Version__: ~1.1.1  
-__Extension__: roc-package-web-app  
-__Custom resolve function__:  No  
-### [koa-accesslog](https://www.npmjs.com/package/koa-accesslog)
-__Version__: ~0.0.2  
-__Extension__: roc-package-web-app  
-__Custom resolve function__:  No  
-### [koa-add-trailing-slashes](https://www.npmjs.com/package/koa-add-trailing-slashes)
-__Version__: ~1.1.0  
-__Extension__: roc-package-web-app  
-__Custom resolve function__:  No  
-### [koa-compressor](https://www.npmjs.com/package/koa-compressor)
-__Version__: ~1.0.3  
-__Extension__: roc-package-web-app  
-__Custom resolve function__:  No  
-### [koa-errors](https://www.npmjs.com/package/koa-errors)
-__Version__: ~1.0.1  
-__Extension__: roc-package-web-app  
-__Custom resolve function__:  No  
-### [koa-etag](https://www.npmjs.com/package/koa-etag)
-__Version__: ~2.0.0  
-__Extension__: roc-package-web-app  
-__Custom resolve function__:  No  
-### [koa-favicon](https://www.npmjs.com/package/koa-favicon)
-__Version__: ~1.2.0  
-__Extension__: roc-package-web-app  
-__Custom resolve function__:  No  
-### [koa-helmet](https://www.npmjs.com/package/koa-helmet)
-__Version__: ~0.3.0  
-__Extension__: roc-package-web-app  
-__Custom resolve function__:  No  
-### [koa-logger](https://www.npmjs.com/package/koa-logger)
-__Version__: ~1.3.0  
-__Extension__: roc-package-web-app  
-__Custom resolve function__:  No  
-### [koa-lowercase-path](https://www.npmjs.com/package/koa-lowercase-path)
-__Version__: ~1.0.0  
-__Extension__: roc-package-web-app  
-__Custom resolve function__:  No  
-### [koa-mount](https://www.npmjs.com/package/koa-mount)
-__Version__: ~1.3.0  
-__Extension__: roc-package-web-app  
-__Custom resolve function__:  No  
-### [koa-normalize-path](https://www.npmjs.com/package/koa-normalize-path)
-__Version__: ~1.0.0  
-__Extension__: roc-package-web-app  
-__Custom resolve function__:  No  
-### [koa-remove-trailing-slashes](https://www.npmjs.com/package/koa-remove-trailing-slashes)
-__Version__: ~1.0.0  
-__Extension__: roc-package-web-app  
-__Custom resolve function__:  No  
-### [koa-static](https://www.npmjs.com/package/koa-static)
-__Version__: ~2.0.0  
-__Extension__: roc-package-web-app  
-__Custom resolve function__:  No  
 
 ## Uses
 ### [koa](https://www.npmjs.com/package/koa)
@@ -73,6 +17,39 @@ __Version__: ~0.0.2
 __Extension__: roc-package-web-app  
 ### [koa-add-trailing-slashes](https://www.npmjs.com/package/koa-add-trailing-slashes)
 __Version__: ~1.1.0  
+__Extension__: roc-package-web-app  
+### [koa-compressor](https://www.npmjs.com/package/koa-compressor)
+__Version__: ~1.0.3  
+__Extension__: roc-package-web-app  
+### [koa-errors](https://www.npmjs.com/package/koa-errors)
+__Version__: ~1.0.1  
+__Extension__: roc-package-web-app  
+### [koa-etag](https://www.npmjs.com/package/koa-etag)
+__Version__: ~2.0.0  
+__Extension__: roc-package-web-app  
+### [koa-favicon](https://www.npmjs.com/package/koa-favicon)
+__Version__: ~1.2.0  
+__Extension__: roc-package-web-app  
+### [koa-helmet](https://www.npmjs.com/package/koa-helmet)
+__Version__: ~0.3.0  
+__Extension__: roc-package-web-app  
+### [koa-logger](https://www.npmjs.com/package/koa-logger)
+__Version__: ~1.3.0  
+__Extension__: roc-package-web-app  
+### [koa-lowercase-path](https://www.npmjs.com/package/koa-lowercase-path)
+__Version__: ~1.0.0  
+__Extension__: roc-package-web-app  
+### [koa-mount](https://www.npmjs.com/package/koa-mount)
+__Version__: ~1.3.0  
+__Extension__: roc-package-web-app  
+### [koa-normalize-path](https://www.npmjs.com/package/koa-normalize-path)
+__Version__: ~1.0.0  
+__Extension__: roc-package-web-app  
+### [koa-remove-trailing-slashes](https://www.npmjs.com/package/koa-remove-trailing-slashes)
+__Version__: ~1.0.0  
+__Extension__: roc-package-web-app  
+### [koa-static](https://www.npmjs.com/package/koa-static)
+__Version__: ~2.0.0  
 __Extension__: roc-package-web-app  
 
 ## Requires
